Add helper function to process double groupedby
This commit introduces a new function that is similar to
transformToDateTableData, with the main difference being its ability to
group data by two attributes. In this case, we use it to group by
Contractor and then by Grade.

In the future, we plan to refactor the function to make it more generic
and improve the code's cleanliness.