Add new component to generate double group by table
This commit adds a new script that was copied from the original
`to-date-table.vue` script. The purpose is to modify it to perform a
double group by on two attributes and display the results in a table
format. This was specifically used to display the double group by table
in the coal production tab on the mining page.

In the future, we plan to refactor this script to make it more flexible
by using a function and allowing it to process multiple group by
attributes.

Notes:
- We have set the style to `height: 90vh` in order to display the table
  without the need for scrolling, as per the design in figma.
- The props have been modified to include primary and secondary slice
  attributes, as well as an attribute header.