Add props to customise yaxis title
Currently, developers are expected to manually input a customised title
for each month line chart, as each page has a specific purpose and
requires a different type of line chart. This does not yet account for
MTD/YTD data, which will be implemented in the future. Currently, we are
keeping the y axis title simple and straightforward so that it is easily
noticeable for the client.